sdk/pyproject.toml
[project]
name = "sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.59.7",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
